requirements.txt
google-genai
openai
PyYAML
python-pptx
python-dotenv
certifi
fastapi
uvicorn
pypdf
python-multipart
